requirements.txt
spotipy
matplotlib
pandas
python-dotenv
tabulate
